# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "liber API",
    "description": "",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "2"
  },
  "paths": {
    "/api/v2/bookmarks": {
      "get": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "list_bookmarks",
        "parameters": [
          {
            "name": "q",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sort",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "folder",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "deep",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged bookmark list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad sort mode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "add_bookmark",
        "requestBody": {
          "description": "Bookmark to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Duplicate accepted idempotently",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBookmark"
                }
              }
            }
          },
          "201": {
            "description": "Bookmark created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBookmark"
                }
              }
            }
          },
          "400": {
            "description": "Missing url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate url",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/settings": {
      "get": {
        "tags": [
          "crate::settings"
        ],
        "operationId": "get_settings",
        "responses": {
          "200": {
            "description": "Settable config values",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "crate::settings"
        ],
        "operationId": "set_setting",
        "requestBody": {
          "description": "Setting as {key, value}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Unknown key or bad backend",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/sync/export": {
      "post": {
        "tags": [
          "crate::sync"
        ],
        "operationId": "export_oplog",
        "requestBody": {
          "description": "Cursor as {since?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Oplog entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/sync/import": {
      "post": {
        "tags": [
          "crate::sync"
        ],
        "operationId": "import_oplog",
        "requestBody": {
          "description": "Bundle as {entries[]}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Merge report counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad bundle",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/sync/prune": {
      "post": {
        "tags": [
          "crate::sync"
        ],
        "operationId": "prune_oplog_ep",
        "requestBody": {
          "description": "Retention as {days?, default 90}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Pruned entry count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "archive": {
            "type": "boolean"
          },
          "confirm_dup": {
            "type": "boolean"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "folder": {
            "type": [
              "string",
              "null"
            ]
          },
          "markdown": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
      "ApiAttachment": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ApiBookmark": {
        "type": "object",
        "required": [
          "uuid",
          "url",
          "title",
          "created_at",
          "updated_at",
          "has_markdown",
          "has_archive"
        ],
        "properties": {
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiAttachment"
            }
          },
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "folder": {
            "type": "string"
          },
          "has_archive": {
            "type": "boolean"
          },
          "has_markdown": {
            "type": "boolean"
          },
          "last_opened_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "open_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "short_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ListResponse": {
        "type": "object",
        "required": [
          "total",
          "page",
          "per_page",
          "bookmarks"
        ],
        "properties": {
          "bookmarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiBookmark"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  }
}